fix: don't duplicate Sol USDC when entering swap from asset details (1st commit)
