+++ app/components/UI/Bridge/hooks/useInitialDestToken/index.ts
@@ -3,11 +3,13 @@ import { useDispatch, useSelector } from 'react-redux';
 import { useEffect } from 'react';
 import { DefaultSwapDestTokens } from '../../constants/default-swap-dest-tokens';
 import { selectChainId } from '../../../../../selectors/networkController';
-import { RouteProp , useRoute } from '@react-navigation/native';
-import { BridgeRouteParams } from '../useInitialSourceToken';
-import { BridgeViewMode } from '../../types';
+import { RouteProp, useRoute } from '@react-navigation/native';
+import { BridgeRouteParams } from '../../Views/BridgeView';
+import { BridgeViewMode, BridgeToken } from '../../types';
 
-export const useInitialDestToken = () => {
+// Need to pass in the initial source token to avoid a race condition with useInitialSourceToken
+// Can't just use selectSourceToken because of race condition
+export const useInitialDestToken = (initialSourceToken: BridgeToken) => {
   const route = useRoute<RouteProp<{ params: BridgeRouteParams }, 'params'>>();
   const dispatch = useDispatch();
   const chainId = useSelector(selectChainId);
@@ -16,8 +18,12 @@ export const useInitialDestToken = () => {
 
   useEffect(() => {
     const defaultDestToken = DefaultSwapDestTokens[chainId];
-    if (isSwap && defaultDestToken) {
+    if (
+      isSwap &&
+      defaultDestToken &&
+      initialSourceToken?.address !== defaultDestToken.address
+    ) {
       dispatch(setDestToken(defaultDestToken));
     }
-  }, [dispatch, chainId, isSwap]);
+  }, [dispatch, chainId, isSwap, initialSourceToken?.address]);
 };
